+++ src/tests/functional/shared_test_classes/src/base/ov_subgraph.cpp
@@ -503,16 +503,16 @@ void SubgraphBaseTest::validate() {
 
 #if 1
     size_t expected_size = expectedOutputs[0].get_size();
-    auto expected_data = static_cast<int *>(expectedOutputs[0].data(ov::element::i32));
+    auto expected_data = static_cast<float *>(expectedOutputs[0].data(ov::element::f32));
     for (size_t i = 0; i < expected_size; i++) {
-        std::cout << static_cast<int>(expected_data[i]) << " ";
+        std::cout << static_cast<float>(expected_data[i]) << " ";
     }
     std::cout << std::endl;
 
     size_t actual_size = actualOutputs[0].get_size();
-    auto actual_data = static_cast<int *>(actualOutputs[0].data(ov::element::i32));
+    auto actual_data = static_cast<float *>(actualOutputs[0].data(ov::element::f32));
     for (size_t i = 0; i < actual_size; i++) {
-        std::cout << static_cast<int>(actual_data[i]) << " ";
+        std::cout << static_cast<float>(actual_data[i]) << " ";
     }
     std::cout << std::endl;
 #endif
